# models/SMS_db/generated/SubjectsModelGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE SubjectsModelGenerated.js PLEASE EDIT ../SubjectsModel.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
// Database
import Database from "../../../classes/Database_SMS_db";
import mongoose, { Schema } from "mongoose";

// Logger
import Logger from "../../../classes/Logger";

const generatedModel = {
  /**
   * Init  schema
   */
  init() {
    const db = Database.getConnection();

    /**
      * Subjects
      */
    const subjectsSchema = new mongoose.Schema({
      averageresultinlastexam: {
        type: "Number"
      },
      class: {
        type: "Number", 
        required: true
      },
      name: {
        type: "String", 
        required: true
      },
      period: {
        type: "Number"
      },
      teacher: {
        type: "String", 
        required: true
      },
      // RELATIONS
      _Class: [{
        type: Schema.ObjectId,
        ref: "Class"
      }],
      _Teachers: [{
        type: Schema.ObjectId,
        ref: "Teachers"
      }],
      
      
      // EXTERNAL RELATIONS
      /*
      _Subjects: [{
        type: Schema.ObjectId,
        ref: "Exams"
      }],
      _Subjects: [{
        type: Schema.ObjectId,
        ref: "Lessons"
      }],
      _Subjects: [{
        type: Schema.ObjectId,
        ref: "Test"
      }],
      */
    });

    generatedModel.setModel(db.connection.model("Subjects", subjectsSchema));

    return subjectsSchema;
  },

  /**
   * Set Model
   */
  setModel: model => {
    generatedModel.model = model;
  },

  /**
   * Get model
   */
  getModel: () => {
    return generatedModel.model;
  },

  // Start queries
    

  // CRUD METHODS


  /**
  * SubjectsModel.create
  *   @description CRUD ACTION create
  *
  */
  async create(item) {
    const obj = new generatedModel.model(item);
    return await obj.save();
  },
  
  /**
  * SubjectsModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *   @param String name Name
  *
  */
  async delete(id) {
    return await generatedModel.model.findByIdAndRemove(id);
  },
  
  /**
  * SubjectsModel.findBy_Class
  *   @description CRUD ACTION findBy_Class
  *   @param Number class Class
  *
  */
  async findBy_Class(key) {
    return await generatedModel.model.find({ '_Class' : key});
  },
  
  /**
  * SubjectsModel.findBy_Teachers
  *   @description CRUD ACTION findBy_Teachers
  *   @param String firstname First Name
  *   @param String lastname Last Name
  *
  */
  async findBy_Teachers(key) {
    return await generatedModel.model.find({ '_Teachers' : key});
  },
  
  /**
  * SubjectsModel.findByclass
  *   @description CRUD ACTION findByclass
  *
  */
  async findByclass(key) {
    return await generatedModel.model.find({ 'class' : key});
  },
  
  /**
  * SubjectsModel.findByname
  *   @description CRUD ACTION findByname
  *
  */
  async findByname(key) {
    return await generatedModel.model.find({ 'name' : key});
  },
  
  /**
  * SubjectsModel.findByteacher
  *   @description CRUD ACTION findByteacher
  *
  */
  async findByteacher(key) {
    return await generatedModel.model.find({ 'teacher' : key});
  },
  
  /**
  * SubjectsModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *   @param String name Name
  *
  */
  async get(id) {
    return await generatedModel.model.findOne({ _id : id });
  },
  
  /**
  * SubjectsModel.get_Class
  *   @description CRUD ACTION get_Class
  *   @param Number class Class
  *
  */
  async get_Class(id) {
    return await generatedModel.model.findOne({ _id : id}).populate("_Class")
  },
  
  /**
  * SubjectsModel.get_Teachers
  *   @description CRUD ACTION get_Teachers
  *
  */
  async get_Teachers(id) {
    return await generatedModel.model.findOne({ _id : id}).populate("_Teachers")
  },
  
  /**
  * SubjectsModel.list
  *   @description CRUD ACTION list
  *
  */
  async list() {
    return await generatedModel.model.find();
  },
  
  /**
  * SubjectsModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *   @param String name Name
  *
  */
  async update(item) { 
    return await generatedModel.model.findOneAndUpdate({ _id: item._id }, item, {'new': true});
  },
  


};

export default generatedModel;