# swagger.yaml
openapi: 3.0.0
info:
  title: SMS
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /class/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Class
      parameters:
        - name: class
          in: path
          description: Class
          required: true
          schema:
            type: number
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Class
      parameters:
        - name: class
          in: path
          description: Class
          required: true
          schema:
            type: number
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    post:
      summary: CRUD ACTION update
      tags:
        - Class
      parameters:
        - name: class
          in: path
          description: Class
          required: true
          schema:
            type: number
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Class"
  /class/findBy_Teachers/{key}:
    get:
      summary: CRUD ACTION findBy_Teachers
      tags:
        - Class
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
...
  /class/findByclassteacher/{key}:
    get:
      summary: CRUD ACTION findByclassteacher
      tags:
        - Class
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
  /class/{id}/get_Teachers:
    get:
      summary: CRUD ACTION get_Teachers
      tags:
        - Class
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  Exams  -----
...
  /exams/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Exams
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: subjectname
          in: path
          description: Subject Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Exams
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
        - name: subjectname
          in: path
          description: Subject Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exams"
    post:
      summary: CRUD ACTION update
      tags:
        - Exams
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: subjectname
          in: path
          description: Subject Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exams"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Exams"
  /exams/findBy_Class/{key}:
    get:
      summary: CRUD ACTION findBy_Class
      tags:
        - Exams
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exams"
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Class:
      properties:
        _id:
          type: string
        class:
          type: number
        classteacher:
          type: string
        noofstudents:
          type: number
        noofsubjects:
          type: number
        roomno:
          type: object
      required:
        -  _id
        - class
        - classteacher
        - noofstudents
        - noofsubjects
        - roomno
    Exams:
      properties:
        _id:
          type: string
        averageresult:
          type: number
        class:
          type: number
        date:
          type: integer
        done:
          type: boolean
        examname:
          type: string
        invigilator1:
          type: string
        invigilator2:
          type: string
        maxmarks:
          type: number
        roomno:
          type: object
        session:
          type: object
        subjectname:
          type: string
        timing:
          type: object
      required:
        -  _id
        - class
        - examname
        - invigilator1
        - invigilator2
        - roomno
        - session
        - subjectname
...